feat(knowledge): rename flow, rename-safe filters, and Langflow overwrite hygiene
Extract document rename UI into KnowledgeRenameDocumentDialog and orchestration into
useKnowledgeRenameDocument so the knowledge page only wires openRenameDialog,
renameInProgress, and dialog props.

Backend: POST /documents/rename via rename_document_chunks_core (partial/idempotent/
resume paths, collision checks, filter sync); initialize effective_doc_id before
the OpenSearch try for safe error payloads.

Frontend: rename mutation; grid passes indexed filename and document_id;
duplicateCheck sends content_hash aligned with backend hash_id; filter panel/
context/normalization persist data_source_refs; search results carry document_id;
source options prefer stable ids; remove debug console.log in filter panel.

Chat: enrich search filters’ source dimension before set_search_filters.

OpenSearch: use conflicts=proceed on filename-based delete_by_query where needed
to avoid 409 version conflicts during replace/re-inge.

Tests: rename query helpers and search filter clauses.

Chore: frontend tsconfig jsx react-jsx and formatting-only tsconfig tweaks.
